packages/plugin-llamaindex/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neocortex-llamaindex"
version = "0.1.0"
description = "LlamaIndex integrations for TinyHumans memory API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "tinyhumansai>=0.5",
    "llama-index-core>=0.10.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
examples = ["llama-index-llms-openai", "python-dotenv>=1.0"]

[tool.setuptools.packages.find]
include = ["neocortex_llamaindex*"]

[tool.ruff]
line-length = 88
target-version = "py39"
